Merge with dev

+++ platform-server-kernel/src/main/java/uno/cod/platform/server/core/service/TeamInvitationService.java
@@ -12,10 +12,11 @@ import uno.cod.platform.server.core.repository.TeamInvitationRepository;
 import uno.cod.platform.server.core.repository.TeamMemberRepository;
 import uno.cod.platform.server.core.repository.TeamRepository;
 import uno.cod.platform.server.core.repository.UserRepository;
+import uno.cod.platform.server.core.service.mail.MailService;
 
+import javax.mail.MessagingException;
 import javax.transaction.Transactional;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -26,33 +27,39 @@ public class TeamInvitationService {
     private final TeamMemberRepository teamMemberRepository;
     private final UserRepository userRepository;
     private final TeamService teamService;
+    private final MailService mailService;
 
     @Autowired
-    public TeamInvitationService(TeamInvitationRepository repository, TeamRepository teamRepository, TeamMemberRepository teamMemberRepository, UserRepository userRepository, TeamService teamService) {
+    public TeamInvitationService(TeamInvitationRepository repository, TeamRepository teamRepository, TeamMemberRepository teamMemberRepository, UserRepository userRepository, TeamService teamService, MailService mailService) {
         this.repository = repository;
         this.teamRepository = teamRepository;
         this.teamMemberRepository = teamMemberRepository;
         this.userRepository = userRepository;
         this.teamService = teamService;
+        this.mailService = mailService;
     }
 
-    public void create(User invitingUser, String usernameToInvite, String canonicalName) {
-        User user = userRepository.findByUsername(usernameToInvite);
+    public void create(User invitingUser, User user, Team team, boolean existingCheck) throws MessagingException {
         if (user == null) {
             throw new CodunoIllegalArgumentException("user.invalid");
         }
-        Team team = teamRepository.findByCanonicalNameAndEnabledTrue(canonicalName);
         if (team == null) {
             throw new CodunoIllegalArgumentException("team.invalid");
         }
         if (team.getInvitedUsers().contains(user)) {
-            throw new CodunoIllegalArgumentException("team.user.already.invited");
+            if (existingCheck) {
+                throw new CodunoIllegalArgumentException("team.user.already.invited");
+            }
+            return;
         }
         TeamUserKey key = new TeamUserKey();
         key.setUser(user);
         key.setTeam(team);
         if (teamMemberRepository.findOne(key) != null) {
-            throw new CodunoIllegalArgumentException("team.has.member");
+            if (existingCheck) {
+                throw new CodunoIllegalArgumentException("team.has.member");
+            }
+            return;
         }
         TeamInvitation invitation = new TeamInvitation();
         invitation.setKey(key);
@@ -62,9 +69,28 @@ public class TeamInvitationService {
         user.addInvitedTeam(team);
         teamRepository.save(team);
         userRepository.save(user);
+
+        Map<String, Object> params = new HashMap<>();
+        params.put("teamCanonicalName", team.getCanonicalName());
+        params.put("teamName", team.getName());
+        params.put("invitedByFull", invitingUser.getFullName().isEmpty() ? invitingUser.getUsername() : invitingUser.getFullName());
+        params.put("invitedBy", invitingUser.getUsername());
+        params.put("nameFull", user.getFullName().isEmpty() ? user.getUsername() : user.getFullName());
+        params.put("name", user.getUsername());
+        mailService.sendMail(user.getFullName(), user.getEmail(), "Team invitation", "team-invitation", params, Locale.ENGLISH);
+    }
+
+    public void create(User invitingUser, String usernameToInvite, String canonicalName) throws MessagingException {
+        User user = userRepository.findByUsername(usernameToInvite);
+        Team team = teamRepository.findByCanonicalNameAndEnabledTrue(canonicalName);
+        create(invitingUser, user, team, true);
     }
 
     public void acceptInvitation(User user, String canonicalName) {
+        acceptInvitation(user, canonicalName, true);
+    }
+
+    public void acceptInvitation(User user, String canonicalName, boolean invitationCheck) {
         Team team = teamRepository.findByCanonicalNameAndEnabledTrue(canonicalName);
         if (team == null) {
             throw new CodunoIllegalArgumentException("team.invalid");
@@ -73,12 +99,14 @@ public class TeamInvitationService {
         TeamUserKey key = new TeamUserKey();
         key.setTeam(team);
         key.setUser(user);
-        TeamInvitation invitation = repository.findByKey(key);
-        if (invitation == null) {
-            throw new CodunoIllegalArgumentException("team.invite.notfound");
+        if (invitationCheck) {
+            TeamInvitation invitation = repository.findByKey(key);
+            if (invitation == null) {
+                throw new CodunoIllegalArgumentException("team.invite.notfound");
+            }
+            repository.delete(invitation);
         }
         teamService.join(user, team);
-        repository.delete(invitation);
     }
 
     public void declineInvitation(User user, String canonicalName) {
